Move spreadsheet module to spreadsheet folder
